Fix running tests in IntelliJ
