+++ diffuse/src/test/kotlin/com/jakewharton/diffuse/FunctionalTest.kt
@@ -1,17 +1,12 @@
 package com.jakewharton.diffuse
 
-import java.io.File
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.io.TempDir
 
-class FunctionalTest {
-
-  @TempDir
-  lateinit var anotherTempDir: File
+internal class FunctionalTest : BaseFunctionalTest() {
 
   @Test
-  internal fun `diffuse diff on stripe`() = runTest(
+  fun `diffuse diff on stripe`() = runTest(
     mode = "apk",
     root = "stripe",
     artifactA = "paymentsheet-example-release-master.apk",
@@ -19,7 +14,7 @@ class FunctionalTest {
   )
 
   @Test
-  internal fun `diffuse diff on obfuscated artifact`() {
+  fun `diffuse diff on obfuscated artifact`() {
     val output = anotherTempDir.resolve("diffuse-out")
     val apkA = loadResource("github-browser/apkA.apk")
     val apkB = loadResource("github-browser/apkB.apk")
@@ -40,14 +35,14 @@ class FunctionalTest {
   }
 
   @Test
-  internal fun `diffuse diff on aar`() = runTest(
+  fun `diffuse diff on aar`() = runTest(
     mode = "aar",
     root = "lazythreeten",
     artifactA = "lazythreetenbp-release.aar",
   )
 
   @Test
-  internal fun `diffuse diff on signed artifact`() = runTest(
+  fun `diffuse diff on signed artifact`() = runTest(
     mode = "apk",
     root = "otwarty-wykop-mobilny",
     artifactA = "app-release.apk",
@@ -55,55 +50,9 @@ class FunctionalTest {
   )
 
   @Test
-  internal fun `diffuse info returns proper size`() = runTest(
+  fun `diffuse info returns proper size`() = runTest(
     mode = "jar",
     root = "diffuse",
     artifactA = "diffuse-unspecified-r8.jar",
   )
-
-  private fun runTest(
-    mode: String,
-    root: String,
-    artifactA: String,
-    artifactB: String = artifactA,
-    mappingA: String? = null,
-    mappingB: String? = mappingA,
-  ) {
-    val diffOutput = anotherTempDir.resolve("diffuse-diff")
-    val infoOutput = anotherTempDir.resolve("diffuse-info")
-    val artifactAResource = loadResource("$root/$artifactA")
-    val artifactBResource = loadResource("$root/$artifactB")
-    val mappingAResource = mappingA?.let { loadResource("$root/$it") }
-    val mappingBResource = mappingB?.let { loadResource("$root/$it") }
-    val expectedDiff = loadResource("$root/diff.txt").let(::File)
-    val expectedInfo = loadResource("$root/info.txt").let(::File)
-
-    if (mappingAResource == null || mappingBResource == null) {
-      main(
-        "diff",
-        "--$mode", artifactAResource, artifactBResource,
-        "--text", diffOutput.path,
-      )
-    } else {
-      main(
-        "diff",
-        "--$mode", artifactAResource, artifactBResource,
-        "--old-mapping", mappingAResource,
-        "--new-mapping", mappingBResource,
-        "--text", diffOutput.path,
-      )
-    }
-    main(
-      "info",
-      "--$mode", artifactBResource,
-      "--text", infoOutput.path,
-    )
-
-    assertThat(diffOutput).hasSameTextualContentAs(expectedDiff)
-    assertThat(infoOutput).hasSameTextualContentAs(expectedInfo)
-  }
 }
-
-private object Loader
-
-fun loadResource(name: String) = Loader::class.java.classLoader.getResource(name).let(::checkNotNull).file
